The StringIndex should be casted to UINT16* but not UINT8*.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@8466 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/PCD/Dxe/Service.c
@@ -786,7 +786,7 @@ SetWorker (
     
     case PCD_TYPE_STRING:
       if (SetPtrTypeSize (TmpTokenNumber, Size)) {
-        CopyMem (StringTable + *((UINT8 *)InternalData), Data, *Size);
+        CopyMem (StringTable + *((UINT16 *)InternalData), Data, *Size);
         Status = EFI_SUCCESS;
       } else {
         Status = EFI_INVALID_PARAMETER;
